+++ src/storage/storage_transformer/performance_metrics.rs
@@ -5,9 +5,10 @@ use crate::{
     metadata::Metadata,
     storage::{
         store_lock::StoreKeyMutex, ListableStorage, ListableStorageTraits, ReadableListableStorage,
-        ReadableStorage, ReadableStorageTraits, ReadableWritableStorage,
-        ReadableWritableStorageTraits, StorageError, StoreKey, StoreKeyRange, StoreKeyStartValue,
-        StoreKeys, StoreKeysPrefixes, StorePrefix, WritableStorage, WritableStorageTraits,
+        ReadableStorage, ReadableStorageTraits, ReadableWritableListableStorage,
+        ReadableWritableStorage, ReadableWritableStorageTraits, StorageError, StoreKey,
+        StoreKeyRange, StoreKeyStartValue, StoreKeys, StoreKeysPrefixes, StorePrefix,
+        WritableStorage, WritableStorageTraits,
     },
 };
 
@@ -15,7 +16,8 @@ use crate::{
 use crate::storage::{
     store_lock::AsyncStoreKeyMutex, AsyncListableStorage, AsyncListableStorageTraits,
     AsyncReadableListableStorage, AsyncReadableStorage, AsyncReadableStorageTraits,
-    AsyncReadableWritableStorageTraits, AsyncWritableStorage, AsyncWritableStorageTraits,
+    AsyncReadableWritableListableStorage, AsyncReadableWritableStorageTraits, AsyncWritableStorage,
+    AsyncWritableStorageTraits,
 };
 
 use std::sync::{
@@ -113,8 +115,14 @@ impl StorageTransformerExtension for PerformanceMetricsStorageTransformer {
         self.create_transformer(storage)
     }
 
+    fn create_readable_writable_listable_transformer(
+        self: Arc<Self>,
+        storage: ReadableWritableListableStorage,
+    ) -> ReadableWritableListableStorage {
+        self.create_transformer(storage)
+    }
+
     #[cfg(feature = "async")]
-    /// Create an asynchronous readable transformer.
     fn create_async_readable_transformer(
         self: Arc<Self>,
         storage: AsyncReadableStorage,
@@ -123,7 +131,6 @@ impl StorageTransformerExtension for PerformanceMetricsStorageTransformer {
     }
 
     #[cfg(feature = "async")]
-    /// Create an asynchronous writable transformer.
     fn create_async_writable_transformer(
         self: Arc<Self>,
         storage: AsyncWritableStorage,
@@ -132,7 +139,6 @@ impl StorageTransformerExtension for PerformanceMetricsStorageTransformer {
     }
 
     #[cfg(feature = "async")]
-    /// Create an asynchronous listable transformer.
     fn create_async_listable_transformer(
         self: Arc<Self>,
         storage: AsyncListableStorage,
@@ -141,11 +147,18 @@ impl StorageTransformerExtension for PerformanceMetricsStorageTransformer {
     }
 
     #[cfg(feature = "async")]
-    /// Create an asynchronous readable and listable transformer.
     fn create_async_readable_listable_transformer(
         self: Arc<Self>,
         storage: AsyncReadableListableStorage,
     ) -> AsyncReadableListableStorage {
+        self.create_transformer(storage)
+    }
+
+    #[cfg(feature = "async")]
+    fn create_async_readable_writable_listable_transformer(
+        self: Arc<Self>,
+        storage: AsyncReadableWritableListableStorage,
+    ) -> AsyncReadableWritableListableStorage {
         self.create_transformer(storage)
     }
 }